+++ src/get_quote.ts
@@ -5,7 +5,7 @@ import type { quote_params_T } from './quote_params_T.js'
 import { quote_path_ } from './quote_path_.js'
 export function get_quote(
 	{ ticker }:quote_params_T, opts:iex_fetch_opts_I = {}
-):Promise<Response&{ json():Promise<Quote> }> {
+):Promise<[Quote, Response]> {
 	return iex_fetch(quote_path_({ ticker }), opts)
 }
 export {