+++ src/get_ref_data_exchange_symbols.ts
@@ -4,7 +4,7 @@ import type { iex_fetch_opts_I } from './iex_fetch_opts_I.js'
 import { ref_data_exchange_symbols_path_ } from './ref_data_exchange_symbols_path_.js'
 export async function get_ref_data_exchange_symbols(
 	params:get_ref_data_exchange_symbols_opts_T, opts:iex_fetch_opts_I
-):Promise<Response&{ json():Promise<FXSymbols> }> {
+):Promise<[FXSymbols, Response]> {
 	return iex_fetch(ref_data_exchange_symbols_path_(params), opts)
 }
 export {