+++ src/get_ref_data_exchanges.ts
@@ -3,7 +3,7 @@ import type { iex_fetch_opts_I } from './iex_fetch_opts_I.js'
 import { ref_data_exchanges_path_ } from './ref_data_exchanges_path_.js'
 export async function get_ref_data_exchanges(
 	opts:iex_fetch_opts_I = {}
-):Promise<Response&{ json():Promise<string[]>}> {
+):Promise<[string[], Response]> {
 	return iex_fetch(ref_data_exchanges_path_(), opts)
 }
 export {