+++ src/ref_data_exchanges_.ts
@@ -1,6 +1,6 @@
 import { get_ref_data_exchanges } from './get_ref_data_exchanges.js'
 import type { iex_fetch_opts_I } from './iex_fetch_opts_I.js'
 export async function ref_data_exchanges_(opts:iex_fetch_opts_I = {}):Promise<string[]|null> {
-	const res = await get_ref_data_exchanges(opts)
-	return res.ok ? res.json() : null
+	const [ref_data_exchanges, res] = await get_ref_data_exchanges(opts)
+	return res.ok ? ref_data_exchanges : null
 }